feat: simpler logs, auto DB setup, and safer sync
- Creates and seeds the database on first run.
- Stores data in ../data as quiz-<port>.db.
- Logs show friendly names (servidor<port>) instead of random IDs.
- Backups auto-copy the database from the primary using the correct port.
- File copy is more reliable (handles OneDrive quirks).
- Non-primary servers reject client connections.

+++ src/main/java/pt/isec/common/model/user/Teacher.java
@@ -18,7 +18,7 @@ public final class Teacher extends User implements Serializable {
         super(name, email, passwordHash);
     }
 
-    public Teacher(Integer id, String name, String email, String passwordHash, LocalDateTime createdAt) {
+    public Teacher(long id, String name, String email, String passwordHash, LocalDateTime createdAt) {
         super(id, name, email, passwordHash, createdAt);
     }
